# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Payment Relay API
  version: 0.3.0
  description: |
    Internal payment relay — systems, wallets, payouts, invoices, reports, and pay page.

    **Auth**: Protected routes require `X-API-Key: sk_live_...` header.

    **Amounts**: All monetary values are in minor units (cents).
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /systems/{id}:
    get:
      tags: [systems]
      summary: Get system by ID
      operationId: getSystem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: System details (no API key in response)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemPublic'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/systems/{id}:
    get:
      tags: [admin]
      summary: Get system detail with wallets and webhooks
      security:
        - AdminTokenAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSystemDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "System API key from POST /systems (format: sk_live_...)"
    AdminTokenAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: Session token from POST /admin/login (Bearer also accepted)
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
...
    SystemPublic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        prefix:
          type: string
        enabled_countries:
          type: array
          items:
            type: string
        webhook_url:
          type: string
          nullable: true
...
    Wallet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        system_id:
          type: string
          format: uuid
        country:
          type: string
        currency:
          type: string
        balance:
          type: integer
          format: int64
...
    WebhookEndpoint:
      type: object
      properties:
        id:
          type: string
          format: uuid
        system_id:
          type: string
          format: uuid
        url:
          type: string
        label:
          type: string
          nullable: true
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AdminSystemDetail:
      type: object
      properties:
        system:
          $ref: '#/components/schemas/SystemPublic'
        wallets:
          type: array
          items:
            $ref: '#/components/schemas/Wallet'
        webhook_endpoints:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEndpoint'